+++ packages/riverpod_generator/test/integration/dependencies.dart
@@ -43,3 +43,12 @@ int transitiveDependencies(TransitiveDependenciesRef ref) => 0;
 
 @Riverpod(dependencies: [dep, family, Dep2])
 int smallTransitiveDependencyCount(SmallTransitiveDependencyCountRef ref) => 0;
+
+@Riverpod(dependencies: [])
+int emptyDependenciesStateless(EmptyDependenciesStatelessRef ref) => 0;
+
+@Riverpod(dependencies: [])
+class EmptyDependenciesStateful extends _$EmptyDependenciesStateful {
+  @override
+  int build() => 0;
+}