+++ packages/riverpod_lint/lib/riverpod_lint.dart
@@ -7,7 +7,7 @@ import 'src/assists/stateful_to_stateless_provider.dart';
 import 'src/assists/stateless_to_stateful_provider.dart';
 import 'src/assists/wrap_with_consumer.dart';
 import 'src/assists/wrap_with_provider_scope.dart';
-import 'src/lints/avoid_manual_providers_as_generated_provider_depenency.dart';
+import 'src/lints/avoid_manual_providers_as_generated_provider_dependency.dart';
 import 'src/lints/generator_class_extends.dart';
 import 'src/lints/missing_provider_scope.dart';
 import 'src/lints/provider_dependencies.dart';