+++ packages/riverpod_analyzer_utils/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased fix
+
+- Handle cascade operators in ref expressions
+- Fix ref.watch/read/... not being decoded inside Notifier methods of generated providers.
+
 ## 0.2.0 - 2023-03-13
 
 - Added `providerForType` TypeChecker for `ProviderFor` annotation